pyproject.toml
[project]
name = "shine2mqtt"
version = "0.1.0"
authors = [{ name = "Bram Harmsen", email = "crazed-lint-kisser@duck.com" }]
description = "A local Growatt server which listens to your Shine Wifi-X datalogger and publishes to MQTT"
readme = "README.md"
requires-python = ">=3.14"
classifiers = [
  "Programming Language :: Python :: 3",
  "Operating System :: OS Independent",
]
license = "GPL-3.0"
license-files = ["LICENSE"]
dependencies = [
  "aiomqtt>=2.5.0",
  "fastapi>=0.128.0",
  "loguru>=0.7.3",
  "pydantic>=2.12.5",
  "pydantic-settings>=2.0.0",
  "pyyaml>=6.0.3",
  "uvicorn>=0.40.0",
]
[dependency-groups]
dev = [
  # Overrides
  # transient dependency of pip-audit, pip 25.03 contains CVE-2026-1703
  "pip>=26.0",
  # regular dependencies
  "pip-audit>=2.10.0",
  "pytest>=9.0.2",
  "pytest-asyncio>=1.3.0",
  "pytest-cov>=7.0.0",
  "ruff>=0.14.14",
  "ty>=0.0.13",
  "import-linter>=2.10",
]

[project.urls]
Homepage = "https://github.com/bremme/shine2mqtt"
Issues = "https://github.com/bremme/shine2mqtt/issues"

[project.scripts]
shine2mqtt = "shine2mqtt.__main__:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

# Tools ################################################################################

# [tool.uv]

[tool.hatch.build.targets.wheel]
packages = ["src/shine2mqtt"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = [
  "--import-mode=importlib",
  "--cov=shine2mqtt",
  "--cov-report=xml",
  "--cov-report=term",
  "-vv",
]

# [tool.pytest.coverage.run]

[tool.ruff]
line-length = 100
target-version = "py314"

[tool.ruff.lint]
# E (pycodestyle), F (pyflakes), W (pycodestyle warnings), I (isort), B (flake8-bugbear), UP (flake8-pytest-style)
select = ["E", "F", "W", "I", "B", "UP"]
ignore = ["E501"]

[tool.ruff.format]

[tool.ty]
